backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
numpy==1.26.3
Pillow==10.2.0
opencv-python-headless==4.10.0.84
cloudinary==1.41.0

typing_extensions>=4.12.0

roboflow
inference-sdk
supervision
python-dotenv==1.0.1
